Replace spree:site with spree:install
