+++ lib/generators/spree/install/templates/app/assets/javascripts/admin/all.js
