+++ lib/generators/spree/install/templates/app/assets/javascripts/store/all.js
