+++ lib/generators/spree/install/templates/app/assets/stylesheets/admin/all.css
