+++ lib/generators/spree/install/templates/app/assets/stylesheets/store/all.css
